# ui-bikes-app/OpenApi/OpenApiSpec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "bikes app specification",
    "description": "this is an OpenApi Spec for our backend",
    "termsOfService": "some terms of service...",
    "contact": {
      "name": "Yassine El Azami El Idrissi",
      "url": "https://github.com/YassineAzamiIdrissi",
      "email": "yassineazami9@gmail.com"
    },
    "license": {
      "name": "license",
      "url": "license..."
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "local environment"
    },
    {
      "url": "some_url",
      "description": "production environment"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/feedbacks": {
      "post": {
        "tags": [
          "feedback"
        ],
        "operationId": "createdFeedback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/bikes": {
      "get": {
        "tags": [
          "bike"
        ],
        "operationId": "getTimeLineBikes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBikeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "bike"
        ],
        "operationId": "saveNewBike",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BikeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/favorites/{bike-id}": {
      "post": {
        "tags": [
          "bike"
        ],
        "operationId": "addBikeToFavorites",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bike"
        ],
        "operationId": "deleteBikeFromFavorites",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/borrow/{bike-id}": {
      "post": {
        "tags": [
          "bike"
        ],
        "operationId": "borrowBike",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/bikes/shareable/{bike-id}": {
      "patch": {
        "tags": [
          "bike"
        ],
        "operationId": "updateShareableStatus",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/return/{bike-id}": {
      "patch": {
        "tags": [
          "bike"
        ],
        "operationId": "returnBike",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/repairing/{bike-id}": {
      "patch": {
        "tags": [
          "bike"
        ],
        "operationId": "updateRepairingStatus",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/picture/{bike-id}": {
      "patch": {
        "tags": [
          "bike"
        ],
        "operationId": "setBikePicture",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/approve/{bike-id}": {
      "patch": {
        "tags": [
          "bike"
        ],
        "operationId": "approveBike",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/feedbacks/bike/{bike-id}": {
      "get": {
        "tags": [
          "feedback"
        ],
        "operationId": "readAllFeedbacksOfBike",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          },
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFeedbackResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/{bike-id}": {
      "get": {
        "tags": [
          "bike"
        ],
        "operationId": "getBikeById",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BikeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/returned": {
      "get": {
        "tags": [
          "bike"
        ],
        "operationId": "findAllReturnedBikesToConnectedUser",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBorrowedBikeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/own-bikes": {
      "get": {
        "tags": [
          "bike"
        ],
        "operationId": "getMyOwnBikes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBikeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/favorites": {
      "get": {
        "tags": [
          "bike"
        ],
        "operationId": "readAllFavoritesOfThiUser",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFavoritesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/favorites/belongs/{bike-id}": {
      "get": {
        "tags": [
          "bike"
        ],
        "operationId": "isBikeInFavorites",
        "parameters": [
          {
            "name": "bike-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/bikes/borrowed": {
      "get": {
        "tags": [
          "bike"
        ],
        "operationId": "getBorrowedBikes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBorrowedBikeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FeedbackRequest": {
        "required": [
          "bikeId",
          "comment",
          "mark"
        ],
        "type": "object",
        "properties": {
          "bikeId": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string"
          },
          "mark": {
            "maximum": 5,
            "minimum": 0,
            "type": "number",
            "format": "double"
          }
        }
      },
      "BikeRequest": {
        "required": [
          "bikeName",
          "description",
          "shareable",
          "techIdentifier"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "bikeName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "techIdentifier": {
            "type": "string"
          },
          "shareable": {
            "type": "boolean"
          }
        }
      },
...
      "FeedbackResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "ownComment": {
            "type": "boolean"
          }
        }
      },
      "PageResponseFeedbackResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "BikeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ownerName": {
            "type": "string"
          },
          "bikeName": {
            "type": "string"
          },
          "techIdentifier": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "picture": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "shareable": {
            "type": "boolean"
          },
          "repairing": {
            "type": "boolean"
          },
          "borrowed": {
            "type": "boolean"
          },
          "favorite": {
            "type": "boolean"
          }
        }
      },
      "PageResponseBikeResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BikeResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "BorrowedBikeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ownerName": {
            "type": "string"
          },
          "borrowedBy": {
            "type": "string"
          },
          "borrowedDate": {
            "type": "string",
            "format": "date-time"
          },
          "bikeName": {
            "type": "string"
          },
          "techIdentifier": {
            "type": "string"
          },
          "returned": {
            "type": "boolean"
          },
          "accepted": {
            "type": "boolean"
          },
          "favorite": {
            "type": "boolean"
          }
        }
      },
      "PageResponseBorrowedBikeResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BorrowedBikeResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "FavoritesResponse": {
        "type": "object",
        "properties": {
          "bikeId": {
            "type": "integer",
            "format": "int32"
          },
          "bikeName": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "bikeTechIdentifier": {
            "type": "string"
          },
          "borrowed": {
            "type": "boolean"
          }
        }
      },
      "PageResponseFavoritesResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FavoritesResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Jwt approach to secure our backend api",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}