# ui-bikes-app/src/app/services/models/borrowed-bike-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BorrowedBikeResponse {
  accepted?: boolean;
  bikeName?: string;
  borrowedBy?: string;
  borrowedDate?: string;
  favorite?: boolean;
  id?: number;
  ownerName?: string;
  returned?: boolean;
  techIdentifier?: string;
}