# ui-bikes-app/src/app/services/models/page-response-borrowed-bike-response.ts
/* tslint:disable */
/* eslint-disable */
import { BorrowedBikeResponse } from '../models/borrowed-bike-response';
export interface PageResponseBorrowedBikeResponse {
  content?: Array<BorrowedBikeResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}